dados normalizados

+++ src/data_processing_scripts/processar_equipes.py
@@ -27,7 +27,7 @@ def processar_dados_equipes():
     
     df_equipes = df_equipes.toDF(*[col_name.strip().replace(' ', '_').lower() for col_name in df_equipes.columns])
     df_equipes = df_equipes.withColumn("ano", col("ano").cast("int")) \
-                           .withColumn("valor", col("valor").cast("double"))
+                           .withColumn("valor", col("valor").cast("int"))
 
     output_path = "src/data/processed/equipes_saude_processed.parquet" 
     df_equipes.write.mode("overwrite").parquet(output_path)
